% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CentEFWR.R
\name{CentEFWR}
\alias{CentEFWR}
\title{CentEFWR:
Calculate customized centiles of estimated fetal weight (EFW) based on WHO, NICHD, PRB/NICHD, Intergrowth 21, Fetal Medicine Foundation, and Hadlock growth charts.}
\usage{
CentEFWR(data, saveToFile = FALSE, filename = "efwCentiles.csv")
}
\arguments{
\item{saveToFile}{: Logical value indicating whether or not to export the data to a  comma-separated values (CSV) file.}

\item{data:}{A data frame with the following columns:

    GA: Gestational Age (days)

    wt: Maternal Weight (kg). May be left blank if not available.

    ht: Maternal Height (cm). May be left blank if not available.

    parity: Number of previous deliveries (Whole number).May be left blank if not available.

    Sex: Fetal sex (Male or Female or left blank).

    EFW: Estimated Fetal Weight (g) Based on Hadlock's formula; If this column is not present in the data frame,
    then the following columns must be provided to calculate EFW internally:

    AC: Abdominal Circumference (mm)

    HC: Head Circumference (mm)

    FL: Femur Diaphysis Length (mm) and/or

    BPD: Biparietal Diameter (mm)}

\item{filename:}{Name of the CSV file; Only required if saveToFile = TRUE.}
}
\value{
The input data frame with additional columns for the estimated fetal weight and it's percentiles according to six growth standards.
}
\description{
This package provides functionality to calculate the customized percentile
 of Estimated Fetal Weight (EFW) based on popular growth standards.
}
\examples{
# Generate sample data with EFW
data=data.frame(GA=c(240,210,266),wt=c(88,92,100),
ht=c(165,165,170),parity=c(0,5,2),
Sex=c("Male","Female",""),Race=c("Black","","Asian"),
EFW=c(2000,2500,2600))
# Compute the EFW centiles for the sample data
 CentEFWR(data)

 # Generate sample data without EFW
data=data.frame(GA=c(200,210,266),wt=c(88,92,100),
ht=c(165,165,170),parity=c(0,5,2),
Sex=c("Male","Female",""),Race=c("Black","","Asian"),
AC=c(220,270,300),
HC=c(250,290,320),
FL=c(55,60,75))

# Compute the EFW centiles for the sample data
 CentEFWR(data)

}
\references{
Tarca, A.L., et al., A new customized fetal growth standard for African American women: the PRB/NICHD Detroit study. Am J Obstet Gynecol, 2018. 218(2s): p. S679-S691.e4.

Grantz, K.L., et al., Fetal growth standards: the NICHD fetal growth study approach in context with INTERGROWTH-21st and the World Health Organization Multicentre Growth Reference Study. Am J Obstet Gynecol, 2018. 218(2s): p. S641-S655.e28.

Nicolaides, K.H., et al., Fetal Medicine Foundation fetal and neonatal population weight charts. Ultrasound in Obstetrics & Gynecology, 2018. 52(1): p. 44-51.

Stirnemann, J., et al., International estimated fetal weight standards of the INTERGROWTH-21(st) Project. Ultrasound Obstet Gynecol, 2017. 49(4): p. 478-486.

Kiserud, T., et al., The World Health Organization Fetal Growth Charts: A Multinational Longitudinal Study of Ultrasound Biometric Measurements and Estimated Fetal Weight. PLoS medicine, 2017. 14(1): p. e1002220-e1002220.

Hadlock, F.P., R.B. Harrist, and J. Martinez-Poyer, In utero analysis of fetal growth: a sonographic weight standard. Radiology, 1991. 181(1): p. 129-33.
}
